Fixing a few more bugs detected in the recording software
Still trying to sort out why sometimes you completley lose the ability to re-record.  Seems like it gets confused sometimes and stops a recording as soon as it starts.  Usually generates an error after this, but the actual underlying cause is still unknown.  Fix is just to refresh the page, and you are fine.  Only seems to happen if you need to do multiple retakes on a specific phrase ( which is the worst time for that bug to show up, as redoing takes is already super frustrating )